extract setupTest util
